front: move some actions from commonConf reducer to the operationalStudies one
Move the actions upsertViaFromSuggestedOP and upsertSeveralViasFromSuggestedOP.
