+++ front/src/modules/pathfinding/components/Itinerary/ModalSuggestedVias.tsx
@@ -11,6 +11,7 @@ import ModalHeaderSNCF from 'common/BootstrapSNCF/ModalSNCF/ModalHeaderSNCF';
 import { useOsrdConfActions, useOsrdConfSelectors } from 'common/osrdContext';
 import { isVia, matchPathStepAndOp } from 'modules/pathfinding/utils';
 import type { SuggestedOP } from 'modules/trainschedule/components/ManageTrainSchedule/types';
+import type { OperationalStudiesConfSliceActions } from 'reducers/osrdconf/operationalStudiesConf';
 import type { PathStep } from 'reducers/osrdconf/types';
 import { useAppDispatch } from 'store';
 import { formatUicToCi } from 'utils/strings';
@@ -21,7 +22,7 @@ type ModalSuggestedViasProps = {
 };
 
 const ModalSuggestedVias = ({ suggestedVias, launchPathfinding }: ModalSuggestedViasProps) => {
-  const { upsertViaFromSuggestedOP } = useOsrdConfActions();
+  const { upsertViaFromSuggestedOP } = useOsrdConfActions() as OperationalStudiesConfSliceActions;
   const { getVias, getDestination, getPathSteps } = useOsrdConfSelectors();
   const dispatch = useAppDispatch();
   const vias = useSelector(getVias());
