+++ front/src/modules/timesStops/TimesStopsInput.tsx
@@ -9,6 +9,7 @@ import { useScenarioContext } from 'applications/operationalStudies/hooks/useSce
 import { useOsrdConfActions } from 'common/osrdContext';
 import { isVia, matchPathStepAndOp } from 'modules/pathfinding/utils';
 import type { SuggestedOP } from 'modules/trainschedule/components/ManageTrainSchedule/types';
+import type { OperationalStudiesConfSliceActions } from 'reducers/osrdconf/operationalStudiesConf';
 import type { PathStep } from 'reducers/osrdconf/types';
 import { useAppDispatch } from 'store';
 
@@ -66,7 +67,8 @@ type TimesStopsInputProps = {
 const TimesStopsInput = ({ allWaypoints, startTime, pathSteps }: TimesStopsInputProps) => {
   const dispatch = useAppDispatch();
   const { t } = useTranslation('timesStops');
-  const { updatePathSteps, upsertSeveralViasFromSuggestedOP } = useOsrdConfActions();
+  const { updatePathSteps, upsertSeveralViasFromSuggestedOP } =
+    useOsrdConfActions() as OperationalStudiesConfSliceActions;
 
   const [rows, setRows] = useState<TimesStopsInputRow[]>([]);
   const { getTrackSectionsByIds, trackSectionsLoading } = useScenarioContext();
