+++ front/src/reducers/osrdconf/operationalStudiesConf/index.ts
@@ -1,5 +1,6 @@
 import { createSlice, type Draft, type PayloadAction } from '@reduxjs/toolkit';
 
+import type { SuggestedOP } from 'modules/trainschedule/components/ManageTrainSchedule/types';
 import type { TrainScheduleWithDetails } from 'modules/trainschedule/components/Timetable/types';
 import computeBasePathStep from 'modules/trainschedule/helpers/computeBasePathStep';
 import { defaultCommonConf, buildCommonConfReducers } from 'reducers/osrdconf/osrdConfCommon';
@@ -8,6 +9,7 @@ import { convertIsoUtcToLocalTime } from 'utils/date';
 import { msToKmh } from 'utils/physics';
 
 import { builPowerRestrictionReducer } from './powerRestrictionReducer';
+import { upsertPathStep } from '../helpers';
 
 export type OperationalStudiesConfState = OsrdConfState;
 
@@ -46,6 +48,22 @@ export const operationalStudiesConfSlice = createSlice({
       state.speedLimitByTag = speedLimitTag || undefined;
       state.powerRestriction = power_restrictions || [];
       state.constraintDistribution = constraint_distribution || 'STANDARD';
+    },
+    // Use this action to transform an op to via from times and stop table or
+    // from the suggested via modal
+    upsertViaFromSuggestedOP(
+      state: Draft<OperationalStudiesConfState>,
+      action: PayloadAction<SuggestedOP>
+    ) {
+      upsertPathStep(state.pathSteps, action.payload);
+    },
+    upsertSeveralViasFromSuggestedOP(
+      state: Draft<OperationalStudiesConfState>,
+      action: PayloadAction<SuggestedOP[]>
+    ) {
+      action.payload.forEach((suggestedOp) => {
+        upsertPathStep(state.pathSteps, suggestedOp);
+      });
     },
   },
 });
