+++ front/src/reducers/osrdconf/operationalStudiesConf/simulationConfReducers.spec.ts
@@ -1,16 +1,24 @@
 import { describe, it, expect } from 'vitest';
 
 import type { LightRollingStockWithLiveries } from 'common/api/osrdEditoastApi';
+import type { SuggestedOP } from 'modules/trainschedule/components/ManageTrainSchedule/types';
 import type { TrainScheduleWithDetails } from 'modules/trainschedule/components/Timetable/types';
-import { operationalStudiesConfSlice } from 'reducers/osrdconf/operationalStudiesConf';
+import {
+  operationalStudiesConfSlice,
+  type OperationalStudiesConfState,
+} from 'reducers/osrdconf/operationalStudiesConf';
 import { defaultCommonConf } from 'reducers/osrdconf/osrdConfCommon';
 import testCommonConfReducers from 'reducers/osrdconf/osrdConfCommon/__tests__/utils';
 import { createStoreWithoutMiddleware } from 'store';
 
-const createStore = () =>
+import commonConfBuilder from '../osrdConfCommon/__tests__/commonConfBuilder';
+import type { PathStep } from '../types';
+
+const createStore = (extraInitialState?: Partial<OperationalStudiesConfState>) =>
   createStoreWithoutMiddleware({
     [operationalStudiesConfSlice.name]: {
       ...defaultCommonConf,
+      ...extraInitialState,
     },
   });
 
@@ -81,6 +89,74 @@ describe('simulationConfReducer', () => {
         },
       ],
       startTime: '2021-01-01T00:00:00+00:00',
+    });
+  });
+
+  describe('should handle upsertViaFromSuggestedOP', () => {
+    const testDataBuilder = commonConfBuilder();
+
+    // For this action, pathfinding has already been made so we know
+    // all steps will have a positionOnPath
+    const pathStepsData = testDataBuilder
+      .buildPathSteps()
+      .map((step, i) => step && { ...step, positionOnPath: i * 100 });
+
+    const [brest, rennes, lemans, paris, strasbourg] = pathStepsData;
+
+    it('should insert a new via if it comes from the suggested vias modal', () => {
+      const pathSteps = [brest, rennes, paris, strasbourg];
+      const store = createStore({ pathSteps });
+
+      const newVia: SuggestedOP = {
+        opId: 'lemans',
+        track: '60ca8dda-6667-11e3-81ff-01f464e0362d',
+        offsetOnTrack: 426.443,
+        positionOnPath: 200,
+        uic: 396002,
+        coordinates: [47.99542250806296, 0.1918181738752042],
+      };
+
+      const insertedVia: PathStep = {
+        id: 'id3', // the id generated by nextId()
+        positionOnPath: 200,
+        uic: 396002,
+        coordinates: [47.99542250806296, 0.1918181738752042],
+      };
+
+      store.dispatch(operationalStudiesConfSlice.actions.upsertViaFromSuggestedOP(newVia));
+      const state = store.getState()[operationalStudiesConfSlice.name];
+      expect(state.pathSteps).toEqual([brest, rennes, insertedVia, paris, strasbourg]);
+    });
+
+    it('should update an existing via if it comes from the "times and step" table and has been added by selecting it on the map', () => {
+      const pathSteps = [brest, rennes, lemans, paris, strasbourg];
+      const store = createStore({ pathSteps });
+
+      const newVia: SuggestedOP = {
+        pathStepId: 'lemans',
+        track: '60ca8dda-6667-11e3-81ff-01f464e0362d',
+        offsetOnTrack: 426.443,
+        positionOnPath: 200,
+        stopFor: 'PT5M',
+        coordinates: [47.99542250806296, 0.1918181738752042],
+      };
+
+      const updatedVia: PathStep = {
+        id: 'lemans',
+        positionOnPath: 200,
+        track: '60ca8dda-6667-11e3-81ff-01f464e0362d',
+        offset: 426.443,
+        stopFor: 'PT5M',
+        coordinates: [47.99542250806296, 0.1918181738752042],
+        arrival: newVia.arrival,
+        locked: newVia.locked,
+        deleted: newVia.deleted,
+        name: newVia.name,
+      };
+
+      store.dispatch(operationalStudiesConfSlice.actions.upsertViaFromSuggestedOP(newVia));
+      const state = store.getState()[operationalStudiesConfSlice.name];
+      expect(state.pathSteps).toEqual([brest, rennes, updatedVia, paris, strasbourg]);
     });
   });
