+++ front/src/reducers/osrdconf/osrdConfCommon/__tests__/utils.ts
@@ -3,12 +3,11 @@ import { describe, beforeEach, it, expect } from 'vitest';
 
 import { StdcmStopTypes } from 'applications/stdcm/types';
 import type { Distribution, Infra, TrainScheduleBase } from 'common/api/osrdEditoastApi';
-import type { SuggestedOP } from 'modules/trainschedule/components/ManageTrainSchedule/types';
 import type { OperationalStudiesConfSlice } from 'reducers/osrdconf/operationalStudiesConf';
 import { defaultCommonConf } from 'reducers/osrdconf/osrdConfCommon';
 import commonConfBuilder from 'reducers/osrdconf/osrdConfCommon/__tests__/commonConfBuilder';
 import type { StdcmConfSlice } from 'reducers/osrdconf/stdcmConf';
-import type { OsrdConfState, PathStep } from 'reducers/osrdconf/types';
+import type { OsrdConfState } from 'reducers/osrdconf/types';
 import { createStoreWithoutMiddleware } from 'store';
 
 function createStore(
@@ -246,76 +245,6 @@ const testCommonConfReducers = (slice: OperationalStudiesConfSlice | StdcmConfSl
     store.dispatch(slice.actions.deleteItinerary());
     const state = store.getState()[slice.name];
     expect(state.pathSteps).toEqual([null, null]);
-  });
-
-  describe('should handle upsertViaFromSuggestedOP', () => {
-    // For this action, pathfinding has already been made so we know
-    // all steps will have a positionOnPath
-    const pathStepsData = testDataBuilder
-      .buildPathSteps()
-      .map((step, i) => step && { ...step, positionOnPath: i * 100 });
-
-    const [brest, rennes, lemans, paris, strasbourg] = pathStepsData;
-
-    it('should insert a new via if it comes from the suggested vias modal', () => {
-      const pathSteps = [brest, rennes, paris, strasbourg];
-      const store = createStore(slice, {
-        pathSteps,
-      });
-
-      const newVia: SuggestedOP = {
-        opId: 'lemans',
-        track: '60ca8dda-6667-11e3-81ff-01f464e0362d',
-        offsetOnTrack: 426.443,
-        positionOnPath: 200,
-        uic: 396002,
-        coordinates: [47.99542250806296, 0.1918181738752042],
-      };
-
-      const insertedVia: PathStep = {
-        id: 'id3', // the id generated by nextId()
-        positionOnPath: 200,
-        uic: 396002,
-        coordinates: [47.99542250806296, 0.1918181738752042],
-      };
-
-      store.dispatch(slice.actions.upsertViaFromSuggestedOP(newVia));
-      const state = store.getState()[slice.name];
-      expect(state.pathSteps).toEqual([brest, rennes, insertedVia, paris, strasbourg]);
-    });
-
-    it('should update an existing via if it comes from the "times and step" table and has been added by selecting it on the map', () => {
-      const pathSteps = [brest, rennes, lemans, paris, strasbourg];
-      const store = createStore(slice, {
-        pathSteps,
-      });
-
-      const newVia: SuggestedOP = {
-        pathStepId: 'lemans',
-        track: '60ca8dda-6667-11e3-81ff-01f464e0362d',
-        offsetOnTrack: 426.443,
-        positionOnPath: 200,
-        stopFor: 'PT5M',
-        coordinates: [47.99542250806296, 0.1918181738752042],
-      };
-
-      const updatedVia: PathStep = {
-        id: 'lemans',
-        positionOnPath: 200,
-        track: '60ca8dda-6667-11e3-81ff-01f464e0362d',
-        offset: 426.443,
-        stopFor: 'PT5M',
-        coordinates: [47.99542250806296, 0.1918181738752042],
-        arrival: newVia.arrival,
-        locked: newVia.locked,
-        deleted: newVia.deleted,
-        name: newVia.name,
-      };
-
-      store.dispatch(slice.actions.upsertViaFromSuggestedOP(newVia));
-      const state = store.getState()[slice.name];
-      expect(state.pathSteps).toEqual([brest, rennes, updatedVia, paris, strasbourg]);
-    });
   });
 
   it('should handle updateRollingStockComfort', () => {
